Split login and sign-up into separate pages, rename Auth directory
